Add tests demonstrating issue with origin of typealiases
Currently, if a typealias defined in a classpath dependency in a package which exists in the source files for the module, then the classpath typealias will end up in signature files. A typealias should only be emitted if it defined in the source files for the project, not if it is from the classpath. This is fixed in the followup.

Bug: 458735166
Test: Adds tests
Change-Id: I66172cc13db29c143d8540ebd292612490c2fdf1