Track subclasses
This change adds tracking of subclasses in `ClassItem`s. This is done in
preparation for the compatibility checks in the following change.

Now, you can get all sub classes of a `ClassItem`, which for a regular
class includes just the classes that inherit directly from it, and for
an interface all the classes that implement it.

Bug: 447143803
Test: `./gradlew`
Change-Id: Idf86db302317929314c643d987e506301dddd1d6