Stop always calling Item.describe(Boolean) in compareSelectableItems(...)
Previously, `compareSelectableItems(...)` would always call
`Item.describe(Boolean)` while that does remove a little but of
duplicated code it does so at the expense of doing lots of unnecessary
work as that method rarely reports issues. This change inlines the
calls so that it is only done when actually reporting an issue.

Bug: 459795258
Test: ./gradlew
Change-Id: I0b403380d92ce9f24838dd5e86813791a88549e6