Update verification-metadata.xml to consistent state
Previously, updating the `gradle/verification-metadata.xml` file was
done on an adhoc basis that added entries which were not needed and
left entries that were no longer used. This change adds scripts
(based on AndroidX scripts) that addresses those issues. They ensure
that gradle is in a consistent state and resets the file back to a
consistent state before updating. That ensures a consistent result
and removes any unneeded entries.

The script was run to update the files before a follow-up change that
adds a new dependency.

Bug: 458362596
Test: ./gradlew
Change-Id: I4d98da73dddcc997d2cc03a8cf87cd5c06c11156