Handle value class types with type parameters
Currently, if a value class has a type parameter used in the inlined value (e.g. `value class ValueClass<T>(val value: T)`) and a property is of the value class type (e.g. `val valueClassUsage: ValueClass<String> = ValueClass("")`), there is an error when trying to search for the accessors of the property, which use the erased inlined value class type. To address this, when inlining the type the variables from the value class type need to be in scope.

Bug: 460409977
Test: CommonValueClassTest -- without this change, the test fails with an IllegalStateException. It is baselined for the snapshot testing because the `getter` and `setter` properties of `PropertyItem` aren't included in the snapshot.
Change-Id: I07b414110855998243bc14a42e996e7be5807bc8