Add separate lexer mode for balanced braces
Previously, the default mode was used for balanced braces and
unbalanced braces. However, that made it difficult to create an inline
tag `{@if ...}` that has its own mode (for processing expressions and
`else`) but which contained nested brace expressions interleaved with
the `else`. That is because while it was easy to switch from the `if`
mode into the default mode it was not possible to switch back to the
`if` mode on the matching `}` as in the default mode not every `}`
matches a `}`.

This change adds a balanced brace mode which is identical to the
default mode except that the `{` and `}` have to be balanced. That
requires each `{` to push the balanced mode and each `}` to pop the
mode. That means that the `if` mode can push the balanced mode for the
`{` that starts the nested brace expression and when the balanced mode
sees the matching `}` it will pop back to the `if` mode.

In order to ensure matching behavior with the default mode it is
necessary to duplicate every rule in the default mode with a
corresponding rule in the balanced mode.

Bug: 429965593
Test: ./gradlew
Change-Id: Ie748bb6d7f597f93089459b717bc2d439d104a0b