Make allowReadingComments=false ignore package javadoc
Previously, `allowReadingComments=false` did not ignore package javadoc
comments as that would have broken package hiding using `/** @hide */`.
However, that is no longer needed so this change stops treating
package documentation differently.

The `JavadocTest` was broken because it reported two new issues that
were previously unreported when `@hide` in `package-info.java` javadoc
hid the package. The test was fixed by updating the issues and the
comments to explain what is happening and why.

Bug: 461768395
Test: ./gradlew
Change-Id: I617d2dbfae95348fbef2ddcc6e70c4c9fdc3d04e