Add CodebaseAssembler.getPackageInfoFromUnderlyingModel(...)
Previously, the only way to provide information from which a package
is constructed was to supply it as a `PackageDoc` in a `PackageDocs`.
That put the onus on the caller to provide that information even though
it often just wants to get a package. This change provides a mechanism
for the `CodebaseAssembler` to provide information directly from the
underlying model.

That information comes from the `package-info.java` file (or in the
case of snapshot will be the `PackageItem` being snapshot). It also
includes the overview file (from `overview.html`). The latter is
included because the `PackageInfo` class is what is passed to
`createPackageItem(...)` and so needs to include everything that a
`PackageDoc` can include. It is also needed for snapshot as it can get
the `overview` from `PackageItem.overviewDocumentation`.

Bug: 461491536
Test: ./gradlew
Change-Id: Ib6ff339d56772c84114f3cc2af920017a6e2b58d