Rename INVALID_HIDE_DOC_TAG to INVALID_BLOCK_TAG_USE and make an error
Originally, `INVALID_HIDE_DOC_TAG` was only used for `@hide` and the
false positives mentioned in the preceding change prevented it from
being treated as an error. It is now used for other tags and has no
false positives so this change renames it to make it more consistent
and makes it an error to prevent new issues being introduced into
Android.

Bug: 462333324
Test: ./gradlew
Change-Id: I4c6d0af34fa7f2ee3fa81bc14fcf4542f1397cbf