Create a fake value when an unknown value is found
Previously, the psi value computation was throwing an error if it was unable to create an annotation value. This is problematic because it blocks users when there is a problem in metalava, not their code. This changes the behavior to instead use a fake value of `error.UnknownValue.ERROR`, which is meant to indicate to users that there is a bug they should report, but not block development.

Bug: 462204247
Test: UastTestBase -- without the update, this test has a runtime error with K2
Change-Id: Ie5e803c2fdad9d4c8d0d9690da9218511164704b