Add isValidPackage(String) in TurbineCodebaseInitialiser
Previously, it used `TurbineElements.getPackageElement(...)` but that
has a couple of issues:
1. It is inefficient as it creates a `TurbinePackageElement` which is
   immediately discarded.
2. It has a bug where it does not work for the unnamed package. The
   unnamed package is found by passing in "" but that looks for the
   top level package called "" which does not exist.

This change adds an `isValidPackage(String)` method which does the same
check as `TurbineElements.getPackageElement(...)` without the bug and
without creating a `TurbinePackageElement`.

Needed for a follow-up change that makes Turbine handling of
`package.html` files more consistent with Psi.

Bug: 461491536
Test: ./gradlew
Change-Id: I958a1f624d5ba7f2c33e4831fc91cc9295eef6fc