[Bug Fix] Add api versions range to correct api version test
What failed:
This test case failed on the git_androidx-studio-integration branch
which uses a different set of prebuilts which are based on git-main.
This test case relies on prebuilts in AOSP which only has SDKs up to
version 35 while git-main has versions up to 36.1.

This results in different test inputs thus different results between the
two branches, hence why the build was failing for git_androidx-studio-integration

The fix:
* Limit the range of values the test focuses on to the maximum value
  that will work in both branches (i.e. 35)
* This test case will be removed entirely due to the depreication of
  --curent-version, thus the fastest way to fix the build was
implemented for this CL

Bug: 458332984
Test: ./gradlew
Change-Id: I976621e0304a94b4e73263c29c9fb440a65d2b50