[BugFix] Remake --api-versions-for-sources bad value test to use mocks
What failed:
This test case failed on the git_androidx-studio-integration branch
which uses a different set of prebuilts which are based on git-main.
This test case relies on prebuilts in AOSP which only has SDKs up to
version 35 while git-main has versions up to 36.1.

This results in different test inputs thus different results between the
two branches, hence why the build was failing for git_androidx-studio-integration

The fix:
* Remove the test that relies on the envrionments prebuilts directory
* Add a new test that mocks the prebuilts file structure, thus making
  the logic independent on the environment

Bug: 458332984
Test: ./gradlew
Change-Id: I5ffaf3caa225966edc4c0f55f377ec652727a187