Refactor finding main module for multiplatform sources
Updates the computation of the main analysis module to split out listing all modules, so that it can be reused in multiplatform codebase creation. Also filters to KaSourceModule to be able to reliably use the `name` property, and since modules from source are the only ones we want to consider anyway.

Test: ./gradlew test
Bug: 407735063
Change-Id: I57dc5135f80cc667f0d2493499f41a055eb272ee