Remove unnecessary dependence on PsiBasedCodebase from ka codebase creation
The KaModuleProcessor used to require a PsiBasedCodebase, but for the most part that is not required.

This switches it to have a DefaultCodebase and an optional PsiBasedCodebase reference, and uses the optional PsiBasedCodebase when needed.

This is to enable creating a DefaultCodebase with KaModuleProcessor for each source set of a multiplatform project. The places the PsiBasedCodebase are needed (documentation creation, field reference value creation) aren't necessary for the multiplatform model since we don't use documentation for any AndroidX metalava runs, and field references can use constant values instead.

Bug: 407735063
Test: ./gradlew test

Change-Id: Ic5dee313d86372df1d62d62532bd0e175fd8c84c