Refactor some model test setup to reuse for multiplatform tests
In preparation for the followup change which adds multiplatform codebase testing, this refactors a few parts of the test model setup to make it easier to reuse.

This updates the CodebaseContext used by tests to have both a Codebase and MultiplatformCodebase, and updates the main test runner to support creating either a Codebase or MultiplatformCodebase, or both.

Bug: 407735062
Test: ./gradlew test
Change-Id: If69077cf8f526627fd7d858f29f4c5bf45aa59c0