Use isValidPackage(String) instead of packageNameFilter
Previously, `gatherPackageJavadoc(...)` took a `packageNameFilter` to
verify that a package was valid before creating a `PackageItem` for it.
That was needed because Psi and Turbine had different logic for
validation package names. That is now abstracted behind the
`isValidPackage(String)` method. So, this change can now replace
`packageNameFilter` with a direct call to `isValidPackage(String)`.

Bug: 461491536
Test: ./gradlew
Change-Id: I6616bbaba9ecba9de32b06874aad88e6217f9165