Prepare for removal of emptyPackageDocumentationFactory()
Currently, `emptyPackageDocumentationFactory()` is used when the
`PackageInfo` object passed to `createPackageItem(...)` has a `null`
`commentFactory` property. Rather than add a special method to handle
this case it will be simpler to just make sure that the `PackageInfo`
objects that are passed to it never have a `null` `commentFactory`
property.

This change prepares for that by implementing it to throw an error when
called so that subclasses can remove their own implementation and it
will fail if they return a `PackageInfo` with a `null` `commentFactory`
property.

Bug: 461491536
Test: ./gradlew
Change-Id: Id2aba572c03ee138900db1189bf70314c3cf715e