Add StubGenerator.Config and create in StubGenerationOptions
Previously, `StubGenerator` just took `Options` and accessed various
fields from that. That tightly coupled its behavior to the command line
options. This change adds a `Config` object that is constructed from
`StubGenerationOptions` and will eventually insulate `StubGenerator`
from those options.

The first property in the new `Config` object is a `StubWriterConfig`
which was previously initialized in `Options`.

Bug: 464226866
Test: ./gradlew
Change-Id: Ic74c21af5023c07794d7857de2c1c4624ced5034