Add compatibility check for annotation retention
Before this change, if an annotation changed its retention
policy incompatibly (e.g. from source to class/binary)
Metalava wouldn't raise any issues. This could be a problem
if a developer just outright changed the retention, or
performed a migration from java to kotlin and didn't
transition the retention correctly.

Bug: 202967265
Test: `./gradlew`
Change-Id: Id58e46473c4a6ba6761acf65c6de022976b9c1d4