Configure unknown flag ApiFlagAction
When building "everything" stubs and signature files in Android Soong
will pass Metalava the definitions of all the flags that it knows about
but treat them as `enabled/read-write` (aka `KEEP`) so that the
`@FlaggedApi` annotations are kept in the signature and stub files.
However, there are some cases where the sources uses flags that are not
known to Soong (aconfig-declaration not available) so Metalava will
treat them as unknown and default to `REVERT`.

This change allows Soong to configure Metalava to treat unknown flags
the same way as known flags, `enabled/read-write` aka `KEEP`.

Bug: 454856811
Test: ./gradlew
Change-Id: Ia36754e8bcc162e9b0b9594ef44aa7acdbb5e072