Limit processing to common and leaf source sets
This skips creating Codebases for any intermediate source sets in a project (e.g. `webMain` when there are `wasmMain` and `jsMain` source sets that depend on it). Running checks and writing signature files for these intermediate source sets doesn't make sense since the source sets they depend on are the final versions of the APIs.

Bug: 407735063
Test: CommonMultiplatformCodebaseTest

Change-Id: I865ba848101e7301252eeab71b1bed4afe8238aa